Fix New Network Replication Issues (#427)
* Network: Reduced transform sync bandwidth and uncapped transform sync


* Fixed some scaling bugs with replication


---------

+++ Polytoria/scripts/datamodel/services/NetworkService.cs
@@ -978,7 +978,7 @@ public sealed partial class NetworkService : Instance
 		[JsonInclude]
 		public string NetID = null!;
 		[JsonInclude]
-		public float[] Value = null!;
+		public byte[] Value = null!;
 	}
 
 	[MemoryPackable]
